Cube size is now restricted by both the width and height of the container.
The way I implemented it created an area around the cube where clicking won't focus the cube. However, I'm not going to fix this, as I don't agree that there should be a concept of focus for the cube. Any keypress should affect the cube, unless we're focused on something like a textbox.