Keep TNoodle version in FMC sheets solves #213
In FMC sheets, Header is avoided, but Footer is now kept. To avoid that TNoodle version collides with the area where the competitor writes the solution, the position of the footer was changed and now it can be controlled with the variable OFFSET_FOOTER.

TNoodle version in FMC sheets

The changes were made to fulfill what was discussed in #231.

Include TNoodle version in FMC sheet

Problem with indentation solved.